Move ISO CAMT053 specific logic to converter::iso_camt053.
